Added cw changes.

+++ nikita/src/main/java/hw02/Task05.java
@@ -16,7 +16,7 @@ public class Task05 {
         double x2;
 
 
-        d = (double) (Math.pow(b , 2) - 4 * a * c);
+        d = Math.pow(b , 2) - 4 * a * c;
         if (d > 0) {
             x1 = -b + Math.sqrt(d) / 2.0 * a;
             x2 = -b - Math.sqrt(d) / 2.0 * a;
